requirements.txt
fastapi
uvicorn[standard]
anthropic
python-docx
google-api-python-client
google-auth
python-multipart
